Refactor list for test folder
