+++ src/test/java/seedu/ddd/logic/commands/list/ListClientCommandTest.java
@@ -1,4 +1,4 @@
-package seedu.ddd.logic.commands;
+package seedu.ddd.logic.commands.list;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
@@ -12,23 +12,27 @@ import static seedu.ddd.testutil.contact.TypicalContacts.FIONA;
 
 import java.util.Arrays;
 import java.util.Collections;
+import java.util.function.Predicate;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import seedu.ddd.logic.commands.list.ListContactCommand;
+import seedu.ddd.logic.commands.Command;
 import seedu.ddd.model.Model;
 import seedu.ddd.model.ModelManager;
 import seedu.ddd.model.UserPrefs;
+import seedu.ddd.model.contact.client.Client;
+import seedu.ddd.model.contact.common.Contact;
 import seedu.ddd.model.contact.common.predicate.ClientTypePredicate;
 import seedu.ddd.model.contact.common.predicate.NameContainsKeywordsPredicate;
 import seedu.ddd.model.contact.common.predicate.VendorTypePredicate;
+import seedu.ddd.model.contact.vendor.Vendor;
 
 
 /**
- * Contains integration tests (interaction with the Model) and unit tests for ListContactCommand.
+ * Contains integration tests (interaction with the Model) and unit tests for ListClientCommand.
  */
-public class ListContactCommandTest {
+public class ListClientCommandTest {
 
     private Model model = new ModelManager(getTypicalAddressBook(), new UserPrefs());
     private Model expectedModel = new ModelManager(getTypicalAddressBook(), new UserPrefs());
@@ -41,25 +45,28 @@ public class ListContactCommandTest {
 
     @Test
     public void execute_listIsNotFiltered_showsSameList() {
-        assertCommandSuccess(new ListContactCommand(Model.PREDICATE_SHOW_ALL_CONTACTS),
-                model, String.format(MESSAGE_CONTACTS_LISTED_OVERVIEW,
-                        expectedModel.getFilteredContactListSize()), expectedModel);
-    }
-
-    @Test
-    public void executeFilterListByClient() {
         expectedModel.updateFilteredContactList(new ClientTypePredicate());
-        assertCommandSuccess(new ListContactCommand(new ClientTypePredicate()),
+        assertCommandSuccess(new ListClientCommand(new ClientTypePredicate()),
                 model, String.format(MESSAGE_CONTACTS_LISTED_OVERVIEW,
                         expectedModel.getFilteredContactListSize()), expectedModel);
     }
 
     @Test
-    public void executeFilterListByVendor() {
-        expectedModel.updateFilteredContactList(new VendorTypePredicate());
-        assertCommandSuccess(new ListContactCommand(new VendorTypePredicate()),
-                model, String.format(MESSAGE_CONTACTS_LISTED_OVERVIEW,
-                        expectedModel.getFilteredContactListSize()), expectedModel);
+    public void execute_clientsAndVendorsInList_containsOnlyClients() {
+        // should contain both clients and vendors initially
+        assertTrue(model.getFilteredContactList().stream().anyMatch((contact) -> contact instanceof Client));
+        assertTrue(model.getFilteredContactList().stream().anyMatch((contact) -> contact instanceof Vendor));
+
+        String expectedMessage = String.format(MESSAGE_CONTACTS_LISTED_OVERVIEW, 4);
+        Predicate<Contact> predicate = new ClientTypePredicate();
+        Command command = new ListVendorCommand(predicate);
+
+        expectedModel.updateFilteredContactList(predicate);
+        assertCommandSuccess(command, model, expectedMessage, expectedModel);
+
+        // should only contain vendors after the command
+        assertTrue(model.getFilteredContactList().stream().anyMatch((contact) -> contact instanceof Client));
+        assertFalse(model.getFilteredContactList().stream().anyMatch((contact) -> contact instanceof Vendor));
     }
 
     @Test
@@ -68,40 +75,40 @@ public class ListContactCommandTest {
                 new NameContainsKeywordsPredicate(Collections.singletonList("first"));
         NameContainsKeywordsPredicate secondPredicate =
                 new NameContainsKeywordsPredicate(Collections.singletonList("second"));
-        ListContactCommand listContactFirstCommand = new ListContactCommand(firstPredicate);
-        ListContactCommand listContactSecondCommand = new ListContactCommand(secondPredicate);
+        ListClientCommand listClientFirstCommand = new ListClientCommand(firstPredicate);
+        ListClientCommand listClientSecondCommand = new ListClientCommand(secondPredicate);
 
         // same object -> return true
-        assertTrue(listContactFirstCommand.equals(listContactFirstCommand));
+        assertTrue(listClientFirstCommand.equals(listClientFirstCommand));
 
         // same values -> return true
-        ListContactCommand listFirstContactCommandCopy = new ListContactCommand(firstPredicate);
-        assertTrue(listContactFirstCommand.equals(listFirstContactCommandCopy));
+        ListClientCommand listFirstClientCommandCopy = new ListClientCommand(firstPredicate);
+        assertTrue(listClientFirstCommand.equals(listFirstClientCommandCopy));
 
         // different types -> return false
-        assertFalse(listContactFirstCommand.equals(1));
+        assertFalse(listClientFirstCommand.equals(1));
 
         // null -> return false
-        assertFalse(listContactFirstCommand.equals(null));
+        assertFalse(listClientFirstCommand.equals(null));
 
         // different person -> return false
-        assertFalse(listContactFirstCommand.equals(listContactSecondCommand));
+        assertFalse(listClientFirstCommand.equals(listClientSecondCommand));
     }
 
     @Test
-    public void execute_zeroKeywords_noContactFound() {
+    public void execute_zeroKeywords_noClientsFound() {
         String expectedMessage = String.format(MESSAGE_CONTACTS_LISTED_OVERVIEW, 0);
         NameContainsKeywordsPredicate predicate = preparePredicate(" ");
-        ListContactCommand command = new ListContactCommand(predicate);
+        ListClientCommand command = new ListClientCommand(predicate);
         expectedModel.updateFilteredContactList(predicate);
         assertCommandSuccess(command, model, expectedMessage, expectedModel);
         assertEquals(Collections.emptyList(), model.getFilteredContactList());
     }
     @Test
-    public void execute_multipleKeywords_multipleContactsFound() {
+    public void execute_multipleKeywords_multipleClientsFound() {
         String expectedMessage = String.format(MESSAGE_CONTACTS_LISTED_OVERVIEW, 3);
         NameContainsKeywordsPredicate predicate = preparePredicate("Kurz Elle Kunz");
-        ListContactCommand command = new ListContactCommand(predicate);
+        ListClientCommand command = new ListClientCommand(predicate);
         expectedModel.updateFilteredContactList(predicate);
         assertCommandSuccess(command, model, expectedMessage, expectedModel);
         assertEquals(Arrays.asList(CARL, ELLE, FIONA), model.getFilteredContactList());
@@ -111,12 +118,12 @@ public class ListContactCommandTest {
         NameContainsKeywordsPredicate predicateOne = new NameContainsKeywordsPredicate(Arrays.asList("keyword"));
         ClientTypePredicate predicateTwo = new ClientTypePredicate();
         VendorTypePredicate predicateThree = new VendorTypePredicate();
-        ListContactCommand listCommandOne = new ListContactCommand(predicateOne);
-        ListContactCommand listCommandTwo = new ListContactCommand(predicateTwo);
-        ListContactCommand listCommandThree = new ListContactCommand(predicateThree);
-        String expectedOne = ListContactCommand.class.getCanonicalName() + "{predicate=" + predicateOne + "}";
-        String expectedTwo = ListContactCommand.class.getCanonicalName() + "{predicate=" + predicateTwo + "}";
-        String expectedThree = ListContactCommand.class.getCanonicalName() + "{predicate=" + predicateThree + "}";
+        ListClientCommand listCommandOne = new ListClientCommand(predicateOne);
+        ListClientCommand listCommandTwo = new ListClientCommand(predicateTwo);
+        ListClientCommand listCommandThree = new ListClientCommand(predicateThree);
+        String expectedOne = ListClientCommand.class.getCanonicalName() + "{predicate=" + predicateOne + "}";
+        String expectedTwo = ListClientCommand.class.getCanonicalName() + "{predicate=" + predicateTwo + "}";
+        String expectedThree = ListClientCommand.class.getCanonicalName() + "{predicate=" + predicateThree + "}";
 
         assertEquals(expectedOne, listCommandOne.toString());
         assertEquals(expectedTwo, listCommandTwo.toString());
