+++ src/test/java/seedu/ddd/logic/commands/list/ListContactCommandTest.java
@@ -1,4 +1,4 @@
-package seedu.ddd.logic.commands;
+package seedu.ddd.logic.commands.list;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
@@ -12,26 +12,22 @@ import static seedu.ddd.testutil.contact.TypicalContacts.FIONA;
 
 import java.util.Arrays;
 import java.util.Collections;
-import java.util.function.Predicate;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import seedu.ddd.logic.commands.list.ListVendorCommand;
 import seedu.ddd.model.Model;
 import seedu.ddd.model.ModelManager;
 import seedu.ddd.model.UserPrefs;
-import seedu.ddd.model.contact.client.Client;
-import seedu.ddd.model.contact.common.Contact;
 import seedu.ddd.model.contact.common.predicate.ClientTypePredicate;
 import seedu.ddd.model.contact.common.predicate.NameContainsKeywordsPredicate;
 import seedu.ddd.model.contact.common.predicate.VendorTypePredicate;
-import seedu.ddd.model.contact.vendor.Vendor;
+
 
 /**
- * Contains integration tests (interaction with the Model) and unit tests for ListVendorCommand.
+ * Contains integration tests (interaction with the Model) and unit tests for ListContactCommand.
  */
-public class ListVendorCommandTest {
+public class ListContactCommandTest {
 
     private Model model = new ModelManager(getTypicalAddressBook(), new UserPrefs());
     private Model expectedModel = new ModelManager(getTypicalAddressBook(), new UserPrefs());
@@ -44,8 +40,23 @@ public class ListVendorCommandTest {
 
     @Test
     public void execute_listIsNotFiltered_showsSameList() {
+        assertCommandSuccess(new ListContactCommand(Model.PREDICATE_SHOW_ALL_CONTACTS),
+                model, String.format(MESSAGE_CONTACTS_LISTED_OVERVIEW,
+                        expectedModel.getFilteredContactListSize()), expectedModel);
+    }
+
+    @Test
+    public void executeFilterListByClient() {
+        expectedModel.updateFilteredContactList(new ClientTypePredicate());
+        assertCommandSuccess(new ListContactCommand(new ClientTypePredicate()),
+                model, String.format(MESSAGE_CONTACTS_LISTED_OVERVIEW,
+                        expectedModel.getFilteredContactListSize()), expectedModel);
+    }
+
+    @Test
+    public void executeFilterListByVendor() {
         expectedModel.updateFilteredContactList(new VendorTypePredicate());
-        assertCommandSuccess(new ListVendorCommand(new VendorTypePredicate()),
+        assertCommandSuccess(new ListContactCommand(new VendorTypePredicate()),
                 model, String.format(MESSAGE_CONTACTS_LISTED_OVERVIEW,
                         expectedModel.getFilteredContactListSize()), expectedModel);
     }
@@ -56,82 +67,61 @@ public class ListVendorCommandTest {
                 new NameContainsKeywordsPredicate(Collections.singletonList("first"));
         NameContainsKeywordsPredicate secondPredicate =
                 new NameContainsKeywordsPredicate(Collections.singletonList("second"));
-        ListVendorCommand listVendorFirstCommand = new ListVendorCommand(firstPredicate);
-        ListVendorCommand listVendorSecondCommand = new ListVendorCommand(secondPredicate);
+        ListContactCommand listContactFirstCommand = new ListContactCommand(firstPredicate);
+        ListContactCommand listContactSecondCommand = new ListContactCommand(secondPredicate);
 
         // same object -> return true
-        assertTrue(listVendorFirstCommand.equals(listVendorFirstCommand));
+        assertTrue(listContactFirstCommand.equals(listContactFirstCommand));
 
         // same values -> return true
-        ListVendorCommand listFirstVendorCommandCopy = new ListVendorCommand(firstPredicate);
-        assertTrue(listVendorFirstCommand.equals(listFirstVendorCommandCopy));
+        ListContactCommand listFirstContactCommandCopy = new ListContactCommand(firstPredicate);
+        assertTrue(listContactFirstCommand.equals(listFirstContactCommandCopy));
 
         // different types -> return false
-        assertFalse(listVendorFirstCommand.equals(1));
+        assertFalse(listContactFirstCommand.equals(1));
 
         // null -> return false
-        assertFalse(listVendorFirstCommand.equals(null));
+        assertFalse(listContactFirstCommand.equals(null));
 
         // different person -> return false
-        assertFalse(listVendorFirstCommand.equals(listVendorSecondCommand));
+        assertFalse(listContactFirstCommand.equals(listContactSecondCommand));
     }
 
     @Test
-    public void execute_zeroKeywords_noVendorFound() {
+    public void execute_zeroKeywords_noContactFound() {
         String expectedMessage = String.format(MESSAGE_CONTACTS_LISTED_OVERVIEW, 0);
         NameContainsKeywordsPredicate predicate = preparePredicate(" ");
-        ListVendorCommand command = new ListVendorCommand(predicate);
+        ListContactCommand command = new ListContactCommand(predicate);
         expectedModel.updateFilteredContactList(predicate);
         assertCommandSuccess(command, model, expectedMessage, expectedModel);
         assertEquals(Collections.emptyList(), model.getFilteredContactList());
     }
-
     @Test
-    public void execute_multipleKeywords_multipleVendorFound() {
+    public void execute_multipleKeywords_multipleContactsFound() {
         String expectedMessage = String.format(MESSAGE_CONTACTS_LISTED_OVERVIEW, 3);
         NameContainsKeywordsPredicate predicate = preparePredicate("Kurz Elle Kunz");
-        ListVendorCommand command = new ListVendorCommand(predicate);
+        ListContactCommand command = new ListContactCommand(predicate);
         expectedModel.updateFilteredContactList(predicate);
         assertCommandSuccess(command, model, expectedMessage, expectedModel);
         assertEquals(Arrays.asList(CARL, ELLE, FIONA), model.getFilteredContactList());
     }
-
-    @Test
-    public void execute_clientsAndVendorsInList_containsOnlyVendors() {
-        // should contain both clients and vendors initially
-        assertTrue(model.getFilteredContactList().stream().anyMatch((contact) -> contact instanceof Client));
-        assertTrue(model.getFilteredContactList().stream().anyMatch((contact) -> contact instanceof Vendor));
-
-        String expectedMessage = String.format(MESSAGE_CONTACTS_LISTED_OVERVIEW, 3);
-        Predicate<Contact> predicate = new VendorTypePredicate();
-        Command command = new ListVendorCommand(predicate);
-
-        expectedModel.updateFilteredContactList(predicate);
-        assertCommandSuccess(command, model, expectedMessage, expectedModel);
-
-        // should only contain vendors after the command
-        assertFalse(model.getFilteredContactList().stream().anyMatch((contact) -> contact instanceof Client));
-        assertTrue(model.getFilteredContactList().stream().anyMatch((contact) -> contact instanceof Vendor));
-    }
-
     @Test
     public void toStringMethod() {
         NameContainsKeywordsPredicate predicateOne = new NameContainsKeywordsPredicate(Arrays.asList("keyword"));
         ClientTypePredicate predicateTwo = new ClientTypePredicate();
         VendorTypePredicate predicateThree = new VendorTypePredicate();
-        ListVendorCommand listCommandOne = new ListVendorCommand(predicateOne);
-        ListVendorCommand listCommandTwo = new ListVendorCommand(predicateTwo);
-        ListVendorCommand listCommandThree = new ListVendorCommand(predicateThree);
-        String expectedOne = ListVendorCommand.class.getCanonicalName() + "{predicate=" + predicateOne + "}";
-        String expectedTwo = ListVendorCommand.class.getCanonicalName() + "{predicate=" + predicateTwo + "}";
-        String expectedThree = ListVendorCommand.class.getCanonicalName() + "{predicate=" + predicateThree + "}";
+        ListContactCommand listCommandOne = new ListContactCommand(predicateOne);
+        ListContactCommand listCommandTwo = new ListContactCommand(predicateTwo);
+        ListContactCommand listCommandThree = new ListContactCommand(predicateThree);
+        String expectedOne = ListContactCommand.class.getCanonicalName() + "{predicate=" + predicateOne + "}";
+        String expectedTwo = ListContactCommand.class.getCanonicalName() + "{predicate=" + predicateTwo + "}";
+        String expectedThree = ListContactCommand.class.getCanonicalName() + "{predicate=" + predicateThree + "}";
 
         assertEquals(expectedOne, listCommandOne.toString());
         assertEquals(expectedTwo, listCommandTwo.toString());
         assertEquals(expectedThree, listCommandThree.toString());
 
     }
-
     private NameContainsKeywordsPredicate preparePredicate(String userInput) {
         return new NameContainsKeywordsPredicate(Arrays.asList(userInput.split("\\s+")));
     }
