+++ src/test/java/seedu/ddd/logic/commands/list/ListVendorCommandTest.java
@@ -1,4 +1,4 @@
-package seedu.ddd.logic.commands;
+package seedu.ddd.logic.commands.list;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
@@ -17,8 +17,7 @@ import java.util.function.Predicate;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import seedu.ddd.logic.commands.list.ListClientCommand;
-import seedu.ddd.logic.commands.list.ListVendorCommand;
+import seedu.ddd.logic.commands.Command;
 import seedu.ddd.model.Model;
 import seedu.ddd.model.ModelManager;
 import seedu.ddd.model.UserPrefs;
@@ -29,11 +28,10 @@ import seedu.ddd.model.contact.common.predicate.NameContainsKeywordsPredicate;
 import seedu.ddd.model.contact.common.predicate.VendorTypePredicate;
 import seedu.ddd.model.contact.vendor.Vendor;
 
-
 /**
- * Contains integration tests (interaction with the Model) and unit tests for ListClientCommand.
+ * Contains integration tests (interaction with the Model) and unit tests for ListVendorCommand.
  */
-public class ListClientCommandTest {
+public class ListVendorCommandTest {
 
     private Model model = new ModelManager(getTypicalAddressBook(), new UserPrefs());
     private Model expectedModel = new ModelManager(getTypicalAddressBook(), new UserPrefs());
@@ -46,28 +44,10 @@ public class ListClientCommandTest {
 
     @Test
     public void execute_listIsNotFiltered_showsSameList() {
-        expectedModel.updateFilteredContactList(new ClientTypePredicate());
-        assertCommandSuccess(new ListClientCommand(new ClientTypePredicate()),
+        expectedModel.updateFilteredContactList(new VendorTypePredicate());
+        assertCommandSuccess(new ListVendorCommand(new VendorTypePredicate()),
                 model, String.format(MESSAGE_CONTACTS_LISTED_OVERVIEW,
                         expectedModel.getFilteredContactListSize()), expectedModel);
-    }
-
-    @Test
-    public void execute_clientsAndVendorsInList_containsOnlyClients() {
-        // should contain both clients and vendors initially
-        assertTrue(model.getFilteredContactList().stream().anyMatch((contact) -> contact instanceof Client));
-        assertTrue(model.getFilteredContactList().stream().anyMatch((contact) -> contact instanceof Vendor));
-
-        String expectedMessage = String.format(MESSAGE_CONTACTS_LISTED_OVERVIEW, 4);
-        Predicate<Contact> predicate = new ClientTypePredicate();
-        Command command = new ListVendorCommand(predicate);
-
-        expectedModel.updateFilteredContactList(predicate);
-        assertCommandSuccess(command, model, expectedMessage, expectedModel);
-
-        // should only contain vendors after the command
-        assertTrue(model.getFilteredContactList().stream().anyMatch((contact) -> contact instanceof Client));
-        assertFalse(model.getFilteredContactList().stream().anyMatch((contact) -> contact instanceof Vendor));
     }
 
     @Test
@@ -76,61 +56,82 @@ public class ListClientCommandTest {
                 new NameContainsKeywordsPredicate(Collections.singletonList("first"));
         NameContainsKeywordsPredicate secondPredicate =
                 new NameContainsKeywordsPredicate(Collections.singletonList("second"));
-        ListClientCommand listClientFirstCommand = new ListClientCommand(firstPredicate);
-        ListClientCommand listClientSecondCommand = new ListClientCommand(secondPredicate);
+        ListVendorCommand listVendorFirstCommand = new ListVendorCommand(firstPredicate);
+        ListVendorCommand listVendorSecondCommand = new ListVendorCommand(secondPredicate);
 
         // same object -> return true
-        assertTrue(listClientFirstCommand.equals(listClientFirstCommand));
+        assertTrue(listVendorFirstCommand.equals(listVendorFirstCommand));
 
         // same values -> return true
-        ListClientCommand listFirstClientCommandCopy = new ListClientCommand(firstPredicate);
-        assertTrue(listClientFirstCommand.equals(listFirstClientCommandCopy));
+        ListVendorCommand listFirstVendorCommandCopy = new ListVendorCommand(firstPredicate);
+        assertTrue(listVendorFirstCommand.equals(listFirstVendorCommandCopy));
 
         // different types -> return false
-        assertFalse(listClientFirstCommand.equals(1));
+        assertFalse(listVendorFirstCommand.equals(1));
 
         // null -> return false
-        assertFalse(listClientFirstCommand.equals(null));
+        assertFalse(listVendorFirstCommand.equals(null));
 
         // different person -> return false
-        assertFalse(listClientFirstCommand.equals(listClientSecondCommand));
+        assertFalse(listVendorFirstCommand.equals(listVendorSecondCommand));
     }
 
     @Test
-    public void execute_zeroKeywords_noClientsFound() {
+    public void execute_zeroKeywords_noVendorFound() {
         String expectedMessage = String.format(MESSAGE_CONTACTS_LISTED_OVERVIEW, 0);
         NameContainsKeywordsPredicate predicate = preparePredicate(" ");
-        ListClientCommand command = new ListClientCommand(predicate);
+        ListVendorCommand command = new ListVendorCommand(predicate);
         expectedModel.updateFilteredContactList(predicate);
         assertCommandSuccess(command, model, expectedMessage, expectedModel);
         assertEquals(Collections.emptyList(), model.getFilteredContactList());
     }
+
     @Test
-    public void execute_multipleKeywords_multipleClientsFound() {
+    public void execute_multipleKeywords_multipleVendorFound() {
         String expectedMessage = String.format(MESSAGE_CONTACTS_LISTED_OVERVIEW, 3);
         NameContainsKeywordsPredicate predicate = preparePredicate("Kurz Elle Kunz");
-        ListClientCommand command = new ListClientCommand(predicate);
+        ListVendorCommand command = new ListVendorCommand(predicate);
         expectedModel.updateFilteredContactList(predicate);
         assertCommandSuccess(command, model, expectedMessage, expectedModel);
         assertEquals(Arrays.asList(CARL, ELLE, FIONA), model.getFilteredContactList());
     }
+
+    @Test
+    public void execute_clientsAndVendorsInList_containsOnlyVendors() {
+        // should contain both clients and vendors initially
+        assertTrue(model.getFilteredContactList().stream().anyMatch((contact) -> contact instanceof Client));
+        assertTrue(model.getFilteredContactList().stream().anyMatch((contact) -> contact instanceof Vendor));
+
+        String expectedMessage = String.format(MESSAGE_CONTACTS_LISTED_OVERVIEW, 3);
+        Predicate<Contact> predicate = new VendorTypePredicate();
+        Command command = new ListVendorCommand(predicate);
+
+        expectedModel.updateFilteredContactList(predicate);
+        assertCommandSuccess(command, model, expectedMessage, expectedModel);
+
+        // should only contain vendors after the command
+        assertFalse(model.getFilteredContactList().stream().anyMatch((contact) -> contact instanceof Client));
+        assertTrue(model.getFilteredContactList().stream().anyMatch((contact) -> contact instanceof Vendor));
+    }
+
     @Test
     public void toStringMethod() {
         NameContainsKeywordsPredicate predicateOne = new NameContainsKeywordsPredicate(Arrays.asList("keyword"));
         ClientTypePredicate predicateTwo = new ClientTypePredicate();
         VendorTypePredicate predicateThree = new VendorTypePredicate();
-        ListClientCommand listCommandOne = new ListClientCommand(predicateOne);
-        ListClientCommand listCommandTwo = new ListClientCommand(predicateTwo);
-        ListClientCommand listCommandThree = new ListClientCommand(predicateThree);
-        String expectedOne = ListClientCommand.class.getCanonicalName() + "{predicate=" + predicateOne + "}";
-        String expectedTwo = ListClientCommand.class.getCanonicalName() + "{predicate=" + predicateTwo + "}";
-        String expectedThree = ListClientCommand.class.getCanonicalName() + "{predicate=" + predicateThree + "}";
+        ListVendorCommand listCommandOne = new ListVendorCommand(predicateOne);
+        ListVendorCommand listCommandTwo = new ListVendorCommand(predicateTwo);
+        ListVendorCommand listCommandThree = new ListVendorCommand(predicateThree);
+        String expectedOne = ListVendorCommand.class.getCanonicalName() + "{predicate=" + predicateOne + "}";
+        String expectedTwo = ListVendorCommand.class.getCanonicalName() + "{predicate=" + predicateTwo + "}";
+        String expectedThree = ListVendorCommand.class.getCanonicalName() + "{predicate=" + predicateThree + "}";
 
         assertEquals(expectedOne, listCommandOne.toString());
         assertEquals(expectedTwo, listCommandTwo.toString());
         assertEquals(expectedThree, listCommandThree.toString());
 
     }
+
     private NameContainsKeywordsPredicate preparePredicate(String userInput) {
         return new NameContainsKeywordsPredicate(Arrays.asList(userInput.split("\\s+")));
     }
