+++ src/test/java/seedu/ddd/logic/parser/list/ListVendorCommandParserTest.java
@@ -1,4 +1,4 @@
-package seedu.ddd.logic.parser;
+package seedu.ddd.logic.parser.list;
 
 import static seedu.ddd.logic.parser.CliSyntax.PREFIX_NAME;
 import static seedu.ddd.logic.parser.CommandParserTestUtil.assertParseSuccess;
@@ -6,13 +6,11 @@ import static seedu.ddd.logic.parser.CommandParserTestUtil.assertParseSuccess;
 import org.junit.jupiter.api.Test;
 
 import seedu.ddd.logic.commands.list.ListContactCommand;
+import seedu.ddd.logic.parser.ArgumentMultimap;
 import seedu.ddd.logic.parser.exceptions.ParseException;
-import seedu.ddd.logic.parser.list.ListContactCommandParser;
 
-
-public class ListContactCommandParserTest {
-
-    private ListContactCommandParser parser = new ListContactCommandParser();
+public class ListVendorCommandParserTest {
+    private ListVendorCommandParser parser = new ListVendorCommandParser();
 
     @Test
     public void parse_validArgs_returnsListContactCommand() throws ParseException {
@@ -27,7 +25,3 @@ public class ListContactCommandParserTest {
         assertParseSuccess(parser, " \n Alice \n \t Bob  \t", expectedListCommand);
     }
 }
-
-
-
-
